Break out misc util functions into separate module #7 #8
